Upgrade dependencies and update root layout and not-found page.
