+++ biome.jsonc
@@ -46,7 +46,7 @@
       }
     }
   },
-  "extends": ["ultracite"],
+  "extends": ["ultracite/core", "ultracite/react", "ultracite/next"],
   "files": {
     "ignoreUnknown": false,
     "includes": [
@@ -59,7 +59,8 @@
       "!.vscode",
       "!.vscode/launch.json",
       "!.next",
-      "!examples"
+      "!examples",
+      "!app/globals.css"
     ]
   }
 }
