+++ package.json
@@ -4,8 +4,8 @@
   "description": "Arcade Documentation",
   "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
+    "dev": "next dev --webpack",
+    "build": "next build --webpack",
     "start": "next start",
     "lint": "pnpm dlx ultracite check",
     "format": "pnpm dlx ultracite fix",
@@ -32,52 +32,52 @@
   "homepage": "https://arcade.dev/",
   "dependencies": {
     "@arcadeai/design-system": "^3.13.0",
-    "@next/third-parties": "15.5.4",
-    "@ory/client": "1.22.5",
+    "@next/third-parties": "16.0.1",
+    "@ory/client": "1.22.7",
     "@theguild/remark-mermaid": "0.3.0",
     "@uidotdev/usehooks": "2.4.1",
-    "lucide-react": "0.545.0",
+    "lucide-react": "0.548.0",
     "mdast-util-to-string": "4.0.0",
-    "motion": "12.23.22",
-    "next": "15.5.4",
+    "motion": "12.23.24",
+    "next": "16.0.1",
     "next-sitemap": "4.2.3",
     "nextra": "4.6.0",
     "nextra-theme-docs": "4.6.0",
-    "posthog-js": "1.273.0",
+    "posthog-js": "1.282.0",
     "react": "19.2.0",
     "react-dom": "19.2.0",
-    "react-hook-form": "7.64.0",
-    "react-syntax-highlighter": "15.6.6",
-    "swagger-ui-react": "^5.29.3",
+    "react-hook-form": "7.65.0",
+    "react-syntax-highlighter": "16.1.0",
+    "swagger-ui-react": "^5.30.0",
     "tailwindcss-animate": "1.0.7",
     "unist-util-visit": "5.0.0",
-    "unist-util-visit-parents": "6.0.1",
+    "unist-util-visit-parents": "6.0.2",
     "zustand": "5.0.8"
   },
   "devDependencies": {
-    "@biomejs/biome": "2.2.5",
-    "@tailwindcss/postcss": "4.1.14",
+    "@biomejs/biome": "2.3.2",
+    "@tailwindcss/postcss": "4.1.16",
     "@types/mdast": "4.0.4",
     "@types/mdx": "2.0.13",
-    "@types/node": "24.7.0",
+    "@types/node": "24.9.2",
     "@types/react": "19.2.2",
-    "@types/react-dom": "19.2.1",
+    "@types/react-dom": "19.2.2",
     "@types/react-syntax-highlighter": "15.5.13",
     "@types/unist": "3.0.3",
-    "commander": "14.0.1",
+    "commander": "14.0.2",
     "fast-glob": "3.3.3",
     "husky": "9.1.7",
-    "lint-staged": "16.2.3",
+    "lint-staged": "16.2.6",
     "next-validate-link": "1.6.3",
-    "openai": "6.2.0",
+    "openai": "6.7.0",
     "ora": "9.0.0",
     "pagefind": "1.4.0",
     "picocolors": "1.1.1",
     "postcss": "8.5.6",
-    "tailwindcss": "4.1.14",
+    "tailwindcss": "4.1.16",
     "typescript": "5.9.3",
-    "ultracite": "5.6.1",
-    "vitest": "3.2.4",
+    "ultracite": "6.1.0",
+    "vitest": "4.0.5",
     "zod": "4.1.12"
   },
   "engines": {
@@ -89,5 +89,5 @@
       "pnpm dlx ultracite fix "
     ]
   },
-  "packageManager": "pnpm@10.11.0+sha512.6540583f41cc5f628eb3d9773ecee802f4f9ef9923cc45b69890fb47991d4b092964694ec3a4f738a420c918a333062c8b925d312f42e4f0c263eb603551f977"
+  "packageManager": "pnpm@10.11.0"
 }
