+++ tsconfig.json
@@ -11,7 +11,7 @@
     "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "jsx": "preserve",
+    "jsx": "react-jsx",
     "incremental": true,
     "plugins": [
       {
@@ -30,7 +30,8 @@
     ".next/types/**/*.ts",
     "app/[lang]/page.mdx",
     "mdx-components.ts",
-    "app/en/page.mdx"
+    "app/en/page.mdx",
+    ".next/dev/types/**/*.ts"
   ],
   "exclude": ["node_modules", "agents"]
 }
